portal_service/requirements.txt
# Agent Configuration Portal Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# File Handling
aiofiles==23.2.1

# Logging
python-json-logger==2.0.7

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
